replace ziplogs flags -mi -t with -a
by default, ziplogs will collect a maximum of two days of log files. a weekly archive and cleanup would remove all seven days but only store the most recent two. changing to `-a` ensures that all logs are retained in the archive, ensuring a weekly archive and cleanup isn't lossy.

this appears to have already been done in the linux scripts.